requirements.txt
Django>=4.1
gunicorn==20.0.4
psycopg2-binary
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
Pillow==9.3.0
